feat(ImmutableLoadEliminating): implement ImmutableLoadEliminating optimization (#124)
Some of the data in the data segment generated by AS is immutable, but
binaryen assumes it is mutable during optimization. As a result, some
optimization passes are missed. This PR passes the immutable data
segment range from AS front-end to binaryen through an additional object
in AsModule, and uses an optimization pass to simplify the load
instruction to const instruction.